add data types to spreadsheet

+++ WeatherForecast.cs
@@ -5,7 +5,7 @@ namespace ExcelWorkshop
     public class WeatherForecast
     {
         [SpreadsheetField(cellName: "TemperatureC")]
-        public int TemperatureC { get; set; }
+        public decimal TemperatureC { get; set; }
 
         [SpreadsheetField(cellName: "Summary", length: 10, startWith: "09")]
         public string Summary { get; set; }
